refactor: Adapter와 관련된 데이터 조작을 Adapter 내부로 옮김
