+++ app/src/main/java/woowacourse/movie/activity/MainActivity.kt
@@ -4,19 +4,13 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.domain.Advertisement
 import woowacourse.movie.domain.AdvertisementMock
-import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.MovieMock
-import woowacourse.movie.domain.advertismentPolicy.AdvertisementPolicy
 import woowacourse.movie.domain.advertismentPolicy.MovieAdvertisementPolicy
 import woowacourse.movie.view.MovieAdapter
 import woowacourse.movie.view.data.MovieListViewData
 import woowacourse.movie.view.data.MovieListViewType
 import woowacourse.movie.view.data.MovieViewData
-import woowacourse.movie.view.data.MovieViewDatas
-import woowacourse.movie.view.mapper.AdvertisementMapper.toView
-import woowacourse.movie.view.mapper.MovieMapper.toView
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -33,7 +27,7 @@ class MainActivity : AppCompatActivity() {
         val movieRecyclerView = findViewById<RecyclerView>(R.id.main_movie_list)
         movieRecyclerView.adapter =
             MovieAdapter(
-                makeMovieListViewData(movies, advertisementDatas, advertisementPolicy),
+                movies, advertisementDatas, advertisementPolicy,
                 ::onClickItem
             )
     }
@@ -46,19 +40,5 @@ class MainActivity : AppCompatActivity() {
                 }
             MovieListViewType.ADVERTISEMENT -> Unit
         }
-    }
-
-    private fun makeMovieListViewData(
-        movie: List<Movie>,
-        advertisement: List<Advertisement>,
-        advertisementPolicy: AdvertisementPolicy
-    ): MovieViewDatas {
-        val result = mutableListOf<MovieListViewData>()
-        for (index in movie.indices) {
-            if (index > 0 && index % advertisementPolicy.movieCount == 0)
-                result.add(advertisement[0].toView())
-            result.add(movie[index].toView())
-        }
-        return MovieViewDatas(result)
     }
 }
