+++ app/src/main/java/woowacourse/movie/view/MovieAdapter.kt
@@ -4,17 +4,27 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
+import woowacourse.movie.domain.Advertisement
+import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.advertismentPolicy.AdvertisementPolicy
 import woowacourse.movie.view.data.AdvertisementViewHolder
 import woowacourse.movie.view.data.MovieInfoViewHolder
 import woowacourse.movie.view.data.MovieListViewData
 import woowacourse.movie.view.data.MovieListViewType
 import woowacourse.movie.view.data.MovieViewData
 import woowacourse.movie.view.data.MovieViewDatas
+import woowacourse.movie.view.mapper.AdvertisementMapper.toView
+import woowacourse.movie.view.mapper.MovieMapper.toView
 
 class MovieAdapter(
-    private val movieViewDatas: MovieViewDatas,
+    movie: List<Movie>,
+    advertisement: List<Advertisement>,
+    advertisementPolicy: AdvertisementPolicy,
     val onClickItem: (data: MovieListViewData) -> Unit
 ) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    private var movieViewDatas: MovieViewDatas =
+        makeMovieListViewData(movie, advertisement, advertisementPolicy)
+
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return when (MovieListViewType.values()[viewType]) {
             MovieListViewType.MOVIE -> MovieInfoViewHolder(
@@ -47,4 +57,28 @@ class MovieAdapter(
 
     override fun getItemId(position: Int): Long = position.toLong()
     override fun getItemCount(): Int = movieViewDatas.value.size
+
+    private fun updateMovieListViewData(
+        movie: List<Movie>,
+        advertisement: List<Advertisement>,
+        advertisementPolicy: AdvertisementPolicy
+    ) {
+        movieViewDatas = makeMovieListViewData(movie, advertisement, advertisementPolicy)
+        notifyDataSetChanged()
+    }
+
+    private fun makeMovieListViewData(
+        movie: List<Movie>,
+        advertisement: List<Advertisement>,
+        advertisementPolicy: AdvertisementPolicy
+    ): MovieViewDatas {
+        return mutableListOf<MovieListViewData>().apply {
+            for (index in movie.indices) {
+                if (index > 0 && index % advertisementPolicy.movieCount == 0) add(advertisement[0].toView())
+                add(movie[index].toView())
+            }
+        }.let {
+            MovieViewDatas(it)
+        }
+    }
 }
